requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary~=2.9.6
firebase-admin
google-generativeai